pyproject.toml
[project]
name = "repodnet"
version = "0.1.0"
description = "Modern async networking library for multiplayer games"
readme = "README.md"
license = "LGPL-3.0-or-later"
license-files = ["COPYING"]
requires-python = ">=3.12"
authors = [
    { name = "Walkercito", email = "walekrcitoliver@gmail.com" },
]
keywords = ["networking", "multiplayer", "gamedev", "async", "msgpack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Networking",
    "Typing :: Typed",
]
dependencies = [
    "msgpack>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Walkercito/repod"
Repository = "https://github.com/Walkercito/repod"
Documentation = "https://github.com/Walkercito/repod/blob/main/docs/DOCS.md"
Issues = "https://github.com/Walkercito/repod/issues"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["repod"]

[tool.ruff]
line-length = 92
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E303"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
force-sort-within-sections = true

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.15.4",
    "ty>=0.0.20",
]
examples = [
    "pygame-ce>=2.5.7",
]

[tool.ty.analysis]
allowed-unresolved-imports = ["pygame.**"]
